+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# enrichplot 1.17.3
+
++ align the dots in `treeplot()` (2022-10-1, Sat)
++ fix a bug in color legend of `treeplot()` (2022-10-1, Sat)
+
 # enrichplot 1.17.2
 
 + `autofacet` to automatically split `barplot` and `dotplot` into several facets (2022-09-06, Tue)